requirements.txt
streamlit
pdfplumber
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
python-dotenv
google-generativeai
sentence-transformers
